Revert "Use `next/after` to increment views"
This reverts commit 6561fd17d6256909c40be6395f45d985a078e056.
