+++ app/_components/views/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./views";
