+++ app/_components/views/views-incrementer.tsx
@@ -1,0 +1,24 @@
+"use client";
+
+import { useEffect } from "react";
+import { incrementViews } from "./actions";
+
+type ViewsIncrementerProps = {
+  slug: string;
+  incrementOnMount: boolean;
+};
+
+const ViewsIncrementer = ({
+  slug,
+  incrementOnMount,
+}: ViewsIncrementerProps) => {
+  useEffect(() => {
+    if (incrementOnMount) {
+      incrementViews(slug);
+    }
+  }, [slug, incrementOnMount]);
+
+  return null;
+};
+
+export default ViewsIncrementer;
